backend/requirements/development.txt
Django>=3.2,<5.0
djangorestframework>=3.14.0,<3.15.0
